Create main.py
Automatic storytelling is a challenging text generation task since it
requires generating long,coherent natural language to describe a sensible
sequence of events.This project aims to build an automatic system that
can tell a story based on given sentences, and the performance of the
system needs to be evaluated and the output should be evaluated how
well they can be perceived by humans. The system consists of two main
components: keyword extraction module and story generation
module.The keyword extraction module is designed to extract keywords
from the input storyand generate a corpus. The thesis choose to use topic
model to extract the keywords of each story. The storytelling generator
module should be able to automatically generate pieces of the story with a
coherent natural language based on the sequence of given events. We wil
build the generation module based on the pre-trained model,GPT-2.
According to evaluation results, the system performs well and have
shown that the stories generated from our system can be well perceived
by humans..